+++ src/index.js
@@ -363,6 +363,9 @@ async function run(request, context) {
       if (params.jobId && !isValidUUIDAnyVersion(params.jobId)) {
         return badRequest('Job Id is invalid. Please provide a valid UUID.');
       }
+      if (params.preflightId && !isValidUUID(params.preflightId)) {
+        return badRequest('Preflight Id is invalid. Please provide a valid UUID.');
+      }
       context.params = params;
       context.request = request;